requirements.txt
numpy>=1.18.1
tensorflow>=2.0.1
streamlit>=0.61.0
Pillow>=7.1.2
